done v 7 add little fixes

+++ src/main/java/com/work/exchangeRate/model/Rate.java
@@ -2,19 +2,18 @@ package com.work.exchangeRate.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
-import java.math.BigDecimal;
 import java.util.HashMap;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class Rate {
 
-    private HashMap<String, BigDecimal> rates;
+    private HashMap<String, Double> rates;
 
-    public HashMap<String, BigDecimal> getRates() {
+    public HashMap<String, Double> getRates() {
         return rates;
     }
 
-    public void setRates(HashMap<String, BigDecimal> rates) {
+    public void setRates(HashMap<String, Double> rates) {
         this.rates = rates;
     }
 }
